+++ pyccel/parser/semantic.py
@@ -2687,6 +2687,11 @@ class SemanticParser(BasicParser):
                 dtype = v.class_type
                 if isinstance(value, Literal) and value is not Nil():
                     value = convert_to_literal(value.python_value, dtype)
+                if isinstance(dtype, InhomogeneousTupleType):
+                    # Raise an error as elements are not yet correctly marked with is_argument.
+                    # This leads to printing errors
+                    errors.report("Inhomogeneous tuples are not yet supported as arguments",
+                            severity='error', symbol=expr)
                 clone_var = v.clone(v.name, is_optional = is_optional, is_argument = True)
                 args.append(FunctionDefArgument(clone_var, bound_argument = bound_argument,
                                         value = value, kwonly = kwonly, annotation = expr.annotation))